requirements.txt
# OCS Tracker Application Dependencies

# Core packages
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7  # PostgreSQL driver
python-multipart==0.0.6  # For form data handling
pyjwt==2.8.0  # For JWT token handling
python-dotenv==1.0.0  # For loading environment variables
jinja2==3.1.2  # For HTML templates
httpx==0.25.0  # For HTTP requests
pytest==7.4.2  # For testing

# Auth packages
cryptography==41.0.3  # Required by PyJWT for certain algorithms
passlib[bcrypt]==1.7.4  # For password hashing if needed

# Frontend/templates
aiofiles==23.2.1  # For static file handling

# For the shared model package
setuptools==68.2.2
wheel==0.41.2

# Azure AD integration
msal==1.24.1  # Microsoft Authentication Library for Python (alternative to direct httpx calls)

# For timezone handling
pytz==2023.3

# For CSV operations
# pandas==2.1.0 - temporarily removed due to installation issues

# Job scheduling (if needed)
apscheduler==3.10.4

# For advanced JSON handling
ujson==5.8.0
